requirements.txt
deta~=1.1.0
fastapi~=0.81.0
python-multipart~=0.0.5
jinja2~=3.1.2
python-dotenv~=0.20.0
